[trello-54]: Moved vuex modules to separated directory.
